apps/slack-integrations-online/pyproject.toml
[project]
name = "slack-integrations-online"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-mongodb>=0.9.0",
    "langchain-openai>=1.1.6",
    "loguru>=0.7.3",
    "mem0ai>=1.0.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymongo>=4.15.5",
    "pyyaml>=6.0.3",
    "slack-sdk>=3.39.0",
]
